% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_filetype}
\alias{assert_filetype}
\title{Asserts that the input is of specified filetype}
\usage{
assert_filetype(filename, type)
}
\arguments{
\item{filename}{Input that is checked for being specified filetype.}

\item{type}{Filetype to check for}
}
\description{
Asserts that the input is of specified filetype
}
