% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_regexp}
\alias{assert_regexp}
\title{Asserts that input string contains pattern}
\usage{
assert_regexp(string, pattern)
}
\arguments{
\item{string}{Input text that is being checked for pattern}

\item{pattern}{Pattern to check for}
}
\description{
Asserts that input string contains pattern
}
