% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_type}
\alias{assert_type}
\title{Asserts that input is of specified type}
\usage{
assert_type(input, type)
}
\arguments{
\item{input}{Input that is being checked for type}

\item{type}{Type to check for}
}
\description{
Asserts that input is of specified type
}
