% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic_check_links.R
\name{check_jekyll_links}
\alias{check_jekyll_links}
\title{Check links on Travis

Supports only Jekyll-based sites for now}
\usage{
check_jekyll_links(site_root = ".", jekyll_port = "4000",
  timeout = 1000, verbose = TRUE, ...)
}
\arguments{
\item{site_root}{path for the site}

\item{jekyll_port}{which port to use}

\item{timeout}{how long to wait for jekyll to sucessfully build the site (in
seconds)}

\item{verbose}{should information about progress of the jekyll build be
displayed?}

\item{...}{additional arguments to be passed to check_links}
}
\description{
Check links on Travis

Supports only Jekyll-based sites for now
}
