% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{check_links}
\alias{check_links}
\title{Check links in your documents}
\usage{
check_links(dir = ".", recursive = TRUE, regexp = "\\\\.html?$",
  glob = NULL, root_dir = dir, ignore_pattern = NULL,
  ignore_tag = NULL, check_external = TRUE, only_with_issues = TRUE,
  raise = c("ok", "warning", "error"), by = c("page", "resource"),
  show_summary = TRUE, checker_options = NULL, ...)
}
\arguments{
\item{dir}{The directory to look for documents}

\item{recursive}{Should sub-folders be searched for documents? (default
`TRUE`).}

\item{regexp}{A regular expression matching the names of the files to check.}

\item{glob}{A wildcard pattern matching the names of the files to check.}

\item{root_dir}{The path for the root of the website. By default, the same
value as `dir`, but can be overriden to use another path (if testing only
some files within the directory structure of the site) or a webserver
address (e.g., `http://localhost:4000`, if testing a Jekyll site).}

\item{ignore_pattern}{A vector of regular expressions matching the path of
the links to ignore in the files (see Details).}

\item{ignore_tag}{A vector of HTML tags to ignore.}

\item{check_external}{Should external links be checked? If `FALSE`, only
local links will be checked.}

\item{only_with_issues}{Should the results include only the broken links
(default) or also the valid links?}

\item{raise}{If set to `warning` or `error`, the function will raise a
warning or an error if broken links are found.}

\item{by}{How should the results of the checks be aggregated?}

\item{show_summary}{Should a summary of the results be printed?}

\item{checker_options}{An optional list that specifies which situations
checker should consider invalid. See \code{\link{checker_options}}.}

\item{...}{additional parameters to be passed to `grep` to match the file
names to check.}
}
\value{
a tibble with the name of the file that includes the link, the link,
  the expanded full path (useful for local/relative links), whether the link
  is valid, and possibly the message/HTTP code returned by the server.
}
\description{
Check links in your documents
}
\details{
Data URI and \code{mailto:} links are not checked.

The \code{by} argument controls how the summary of the results is being
displayed. Using \code{page} is typically more convenient for small sites,
while \code{resource} works better for larger websites that use templates
are more likely to have mispecified resources across many pages.

`checker` attempts to respect the rules specified by the robots.txt files
provided by the external servers. Thus some URLs might not be checked.

## Ignore pattern

  If more than one regular expressions is provided to `ignore_pattern`, they
  will be evaluated in succession: thus, the order in which you provide them
  may matter. For local files, matching of the regular expressions is done
  on the fully expanded link paths. Make sure your regular expression don't
  inadvertently match patterns higher up in your tree files to the risk of
  excluding all files from being checked.
}
