% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkeR.R
\name{checkeR}
\alias{checkeR}
\title{Check the version of R}
\usage{
checkeR()
}
\description{
Fetches the latest (not development!) R version and compares it with your currently installed R version (the version of the R session from which you are running this function).
This is a wrapper for the RStudio add-in for \code{check.for.updates.R()} in \code{{installr}}.
}
