% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkeR.R
\name{checkeRStudio}
\alias{checkeRStudio}
\title{Check the version of RStudio}
\usage{
checkeRStudio(GUI = TRUE)
}
\arguments{
\item{GUI}{a logical indicating whether a graphics menu should be used.}
}
\description{
Fetches the latest (not development!) RStudio version and compares it with your currently installed RStudio version (the version of the RStudio session from which you are running this function).
}
