% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_news.R
\name{view_news_RStudio}
\alias{view_news_RStudio}
\title{View NEWS of RStudio}
\usage{
view_news_RStudio()
}
\description{
Function to display the RStudio Release Notes web page in a web browser.
The default web browser will launch and display Posit's RStudio release notes web page.
}
\note{
It just open Posit's RStudio release notes web page with \code{browseURL()}.
}
