% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_regexp}
\alias{assert_regexp}
\title{Asserts that input contains pattern}
\usage{
assert_regexp(input, pattern)
}
\arguments{
\item{pattern}{Pattern to check for}

\item{string}{Input text that is being checked for pattern}
}
\description{
Asserts that input contains pattern
}
