% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{check_links}
\alias{check_links}
\title{Check links in your documents

Currently only HTML files are supported.}
\usage{
check_links(dir = ".", recursive = TRUE, regexp = "\\\\.html?$",
  glob = NULL, only_with_issues = TRUE, show_full_path = FALSE,
  raise = c("ok", "warning", "error"), by = c("page", "resource"),
  show_summary = TRUE, ...)
}
\arguments{
\item{dir}{The directory to look for documents}

\item{recursive}{Should sub-folders be searched for documents? (default
`TRUE`).}

\item{regexp}{A regular expression matching the names of the files to check.}

\item{glob}{A wildcard pattern matching the names fo the files to check.}

\item{only_with_issues}{Should the results include only the broken links
(default) or also the valid links?}

\item{raise}{If set to `warning` or `error`, the function will raise a
warning or an error if broken links are found.}

\item{by}{How should the results of the checks be aggregated?}

\item{show_summary}{Should a summary of the results be printed?}

\item{...}{additional parameters to be passed to `grep` to match the file
names to check.}
}
\value{
a tibble with the name of the file that includes the link, the link,
  the expanded full path (useful for local/relative links), whether the link
  is valid, and possibly the message/HTTP code returned by the server.
}
\description{
Check links in your documents

Currently only HTML files are supported.
}
\details{
Data URI and \code{mailto:} links are not checked.

The \code{by} argument controls how the summary of the results is being
displayed. Using \code{page} is typically more convenient for small sites,
while \code{resource} works better for larger websites that use templates
are more likely to have mispecified resources across many pages.
}
