% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_options.R
\name{checker_options}
\alias{checker_options}
\title{Checker Options}
\usage{
checker_options(checker_options, ...)
}
\arguments{
\item{checker_options}{a named list of the checker options to customize.}

\item{...}{ignored}
}
\value{
a customized list of checker options
}
\description{
Checker Options
}
\details{
Customize which HTTP status codes and other situations lead checker
  to mark links and other URIs as invalid/broken.

Options set to TRUE will be interpreted by checker as valid links. For
instance, you will most likely always want `status_code_404 = FALSE`, so
checker will report as invalid 404 HTTP status codes.

You can customize default values by:

* specifying options in your .Rprofile

* passing a list of the options you want to customize to `check_link()`
using the `checker_options` argument
}
\examples{
## check_links(..., checker_options = list(status_code_300 = TRUE, status_code_301 = TRUE))
}
