% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{chk_requirements}
\alias{chk_requirements}
\title{Check R setup before a class}
\usage{
chk_requirements(path = "inst/default.yaml")
}
\arguments{
\item{path}{path to yaml file}
}
\description{
Check R setup before a class
}
